Flight Price Prediction/requirements.txt
flask
numpy
pandas
scikit-learn
matplotlib
gunicorn
apache-airflow
mlflow